+++ apps/backoffice-v2/src/components/organisms/Subject/Subject.Info.tsx
@@ -1,18 +1,18 @@
-import { FunctionComponent, useCallback, useState } from 'react';
+import { isNullish } from '@ballerine/common';
 import { WarningAlert } from 'components/atoms/WarningAlert';
-import { isValidDate } from '../../../utils/is-valid-date';
-import { formatDate } from '../../../utils/format-date';
 import { DataField } from 'components/molecules/DataField/DataField';
-import { State } from '../../../enums';
-import { IInfoProps } from 'components/organisms/Subject/interfaces';
 import { DetailsGrid } from 'components/molecules/DetailsGrid/DetailsGrid';
-import { toStartCase } from '../../../utils/to-start-case/to-start-case';
+import { Modal } from 'components/organisms/Modal/Modal';
+import { IInfoProps } from 'components/organisms/Subject/interfaces';
+import { useToggle } from 'hooks/useToggle/useToggle';
+import { FunctionComponent, useCallback, useState } from 'react';
+import { State } from '../../../enums';
 import { camelCaseToSpace } from '../../../utils/camel-case-to-space/camel-case-to-space';
-import { isNullish } from '../../../utils/is-nullish/is-nullish';
 import { createArrayOfNumbers } from '../../../utils/create-array-of-numbers/create-array-of-numbers';
 import { ctw } from '../../../utils/ctw/ctw';
-import { Modal } from 'components/organisms/Modal/Modal';
-import { useToggle } from 'hooks/useToggle/useToggle';
+import { formatDate } from '../../../utils/format-date';
+import { isValidDate } from '../../../utils/is-valid-date';
+import { toStartCase } from '../../../utils/to-start-case/to-start-case';
 
 export const useInfo = ({
   whitelist = [],