+++ apps/backoffice-v2/src/lib/react-query/query-client.ts
@@ -1,9 +1,8 @@
+import { isErrorWithMessage, isObject } from '@ballerine/common';
 import { MutationCache, QueryCache, QueryClient } from '@tanstack/react-query';
-import toast from 'react-hot-toast';
 import { t } from 'i18next';
+import toast from 'react-hot-toast';
 import { IGlobalToastContext } from '../../interfaces';
-import { isObject } from '../../utils/is-object/is-object';
-import { isErrorWithMessage } from '../../utils/is-error-with-message/is-error-with-message';
 import { isZodError } from '../../utils/is-zod-error/is-zod-error';
 
 // TODO: Add i18n plurals