+++ sdks/workflow-node-sdk/src/index.ts
@@ -1,7 +1,2 @@
-import { createNodeWorkflow, NodeWorkflowOptions, NodeWorkflow } from "./lib/node-workflow";
-
-type TInitNodeWorkflow = (options: NodeWorkflowOptions) => NodeWorkflow;
-
-export const initNodeWorkflow: TInitNodeWorkflow  = (options)  => {
-  return createNodeWorkflow(options);
-}
+export { createWorkflow, WorkflowNodeSDK } from './lib';
+export type { WorkflowOptionsNode } from './lib';